Add option to estimate the size of vmcore dump files
A user might want to know how much space a vmcore file will take on
the system and how much space on their disk should be available to
save it during a crash.

The option --vmcore-size does not create the vmcore file but provides
an estimation of the size of the final vmcore file created with the
same make dumpfile options.

Signed-off-by: Julien Thierry <jthierry@redhat.com>
Cc: Kazuhito Hagio <k-hagio-ab@nec.com>